moves ducks utils to ducks/index.js
