+++ src/utils/index.js
@@ -1,3 +1,12 @@
 export * from "./api";
 export * from "./helpers";
-export * from "./ducks";
+
+export const mapObj = (obj, fn) =>
+  Object.keys(obj).reduce(
+    (state, itemKey) => ({ ...state, [itemKey]: fn(obj[itemKey]) }),
+    {}
+  );
+
+export const augmentSelectorWith = parentSelector => selector => {
+  return (state, ...restArgs) => selector(parentSelector(state), ...restArgs);
+};
